Reinstate PostgreSQLPlatform and PostgreSQLKeywords
