+++ docs/en/reference/platforms.rst
@@ -55,7 +55,7 @@ Microsoft SQL Server
 PostgreSQL
 ^^^^^^^^^^
 
--  ``PostgreSQL94Platform`` for version 9.4 and above.
+-  ``PostgreSQLPlatform`` for version 9.4 and above.
 -  ``PostgreSQL100Platform`` for version 10.0 and above.
 
 SQLite
