+++ psalm.xml.dist
@@ -61,6 +61,11 @@
                 -->
                 <referencedClass name="Doctrine\DBAL\Driver\ServerInfoAwareConnection"/>
                 <referencedClass name="Doctrine\DBAL\VersionAwarePlatformDriver"/>
+                <!--
+                    TODO: remove in 4.0.0
+                -->
+                <referencedClass name="Doctrine\DBAL\Platforms\Keywords\PostgreSQL94Keywords"/>
+                <referencedClass name="Doctrine\DBAL\Platforms\PostgreSQL94Platform"/>
             </errorLevel>
         </DeprecatedClass>
         <DeprecatedInterface>
