+++ src/Portability/OptimizeFlags.php
@@ -7,7 +7,7 @@ namespace Doctrine\DBAL\Portability;
 use Doctrine\DBAL\Platforms\AbstractPlatform;
 use Doctrine\DBAL\Platforms\DB2Platform;
 use Doctrine\DBAL\Platforms\OraclePlatform;
-use Doctrine\DBAL\Platforms\PostgreSQL94Platform;
+use Doctrine\DBAL\Platforms\PostgreSQLPlatform;
 use Doctrine\DBAL\Platforms\SqlitePlatform;
 use Doctrine\DBAL\Platforms\SQLServer2012Platform;
 
@@ -22,7 +22,7 @@ final class OptimizeFlags
     private static $platforms = [
         DB2Platform::class           => 0,
         OraclePlatform::class        => Connection::PORTABILITY_EMPTY_TO_NULL,
-        PostgreSQL94Platform::class  => 0,
+        PostgreSQLPlatform::class    => 0,
         SqlitePlatform::class        => 0,
         SQLServer2012Platform::class => 0,
     ];
