+++ tests/Driver/AbstractPostgreSQLDriverTest.php
@@ -10,11 +10,12 @@ use Doctrine\DBAL\Driver\API\PostgreSQL;
 use Doctrine\DBAL\Platforms\AbstractPlatform;
 use Doctrine\DBAL\Platforms\PostgreSQL100Platform;
 use Doctrine\DBAL\Platforms\PostgreSQL94Platform;
+use Doctrine\DBAL\Platforms\PostgreSQLPlatform;
 use Doctrine\DBAL\Schema\AbstractSchemaManager;
 use Doctrine\DBAL\Schema\PostgreSQLSchemaManager;
 
 /**
- * @extends AbstractDriverTest<PostgreSQL94Platform>
+ * @extends AbstractDriverTest<PostgreSQLPlatform>
  */
 class AbstractPostgreSQLDriverTest extends AbstractDriverTest
 {
