+++ tests/Functional/Ticket/DBAL168Test.php
@@ -2,6 +2,7 @@
 
 namespace Doctrine\DBAL\Tests\Functional\Ticket;
 
+use Doctrine\DBAL\Platforms\PostgreSQLPlatform;
 use Doctrine\DBAL\Schema\Table;
 use Doctrine\DBAL\Tests\FunctionalTestCase;
 
@@ -9,7 +10,7 @@ class DBAL168Test extends FunctionalTestCase
 {
     public function testDomainsTable(): void
     {
-        if ($this->connection->getDatabasePlatform()->getName() !== 'postgresql') {
+        if ($this->connection->getDatabasePlatform() instanceof PostgreSQLPlatform) {
             self::markTestSkipped('PostgreSQL only test');
         }
