+++ tests/Platforms/AbstractPostgreSQLPlatformTestCase.php
@@ -2,7 +2,7 @@
 
 namespace Doctrine\DBAL\Tests\Platforms;
 
-use Doctrine\DBAL\Platforms\PostgreSQL94Platform;
+use Doctrine\DBAL\Platforms\PostgreSQLPlatform;
 use Doctrine\DBAL\Schema\Column;
 use Doctrine\DBAL\Schema\ColumnDiff;
 use Doctrine\DBAL\Schema\Comparator;
@@ -12,12 +12,13 @@ use Doctrine\DBAL\Schema\Table;
 use Doctrine\DBAL\Schema\TableDiff;
 use Doctrine\DBAL\TransactionIsolationLevel;
 use Doctrine\DBAL\Types\Type;
+use Doctrine\DBAL\Types\Types;
 use UnexpectedValueException;
 
 use function sprintf;
 
 /**
- * @template T of PostgreSQL94Platform
+ * @template T of PostgreSQLPlatform
  * @extends AbstractPlatformTestCase<T>
  */
 abstract class AbstractPostgreSQLPlatformTestCase extends AbstractPlatformTestCase
@@ -983,5 +984,86 @@ abstract class AbstractPostgreSQLPlatformTestCase extends AbstractPlatformTestCa
             "'Foo''Bar\\'",
             $this->platform->getListTableColumnsSQL("Foo'Bar\\.baz_table")
         );
+    }
+
+    public function testSupportsPartialIndexes(): void
+    {
+        self::assertTrue($this->platform->supportsPartialIndexes());
+    }
+
+    public function testGetCreateTableSQLWithUniqueConstraints(): void
+    {
+        $table = new Table('foo');
+        $table->addColumn('id', 'string');
+        $table->addUniqueConstraint(['id'], 'test_unique_constraint');
+        self::assertSame(
+            [
+                'CREATE TABLE foo (id VARCHAR(255) NOT NULL)',
+                'ALTER TABLE foo ADD CONSTRAINT test_unique_constraint UNIQUE (id)',
+            ],
+            $this->platform->getCreateTableSQL($table),
+            'Unique constraints are added to table.'
+        );
+    }
+
+    public function testGetCreateTableSQLWithColumnCollation(): void
+    {
+        $table = new Table('foo');
+        $table->addColumn('id', 'string');
+        $table->addOption('comment', 'foo');
+        self::assertSame(
+            [
+                'CREATE TABLE foo (id VARCHAR(255) NOT NULL)',
+                "COMMENT ON TABLE foo IS 'foo'",
+            ],
+            $this->platform->getCreateTableSQL($table),
+            'Comments are added to table.'
+        );
+    }
+
+    public function testColumnCollationDeclarationSQL(): void
+    {
+        self::assertEquals(
+            'COLLATE "en_US.UTF-8"',
+            $this->platform->getColumnCollationDeclarationSQL('en_US.UTF-8')
+        );
+    }
+
+    public function testHasNativeJsonType(): void
+    {
+        self::assertTrue($this->platform->hasNativeJsonType());
+    }
+
+    public function testReturnsJsonTypeDeclarationSQL(): void
+    {
+        self::assertSame('JSON', $this->platform->getJsonTypeDeclarationSQL([]));
+        self::assertSame('JSON', $this->platform->getJsonTypeDeclarationSQL(['jsonb' => false]));
+        self::assertSame('JSONB', $this->platform->getJsonTypeDeclarationSQL(['jsonb' => true]));
+    }
+
+    public function testReturnsSmallIntTypeDeclarationSQL(): void
+    {
+        self::assertSame(
+            'SMALLSERIAL',
+            $this->platform->getSmallIntTypeDeclarationSQL(['autoincrement' => true])
+        );
+
+        self::assertSame(
+            'SMALLINT',
+            $this->platform->getSmallIntTypeDeclarationSQL(['autoincrement' => false])
+        );
+
+        self::assertSame(
+            'SMALLINT',
+            $this->platform->getSmallIntTypeDeclarationSQL([])
+        );
+    }
+
+    public function testInitializesJsonTypeMapping(): void
+    {
+        self::assertTrue($this->platform->hasDoctrineTypeMappingFor('json'));
+        self::assertEquals(Types::JSON, $this->platform->getDoctrineTypeMapping('json'));
+        self::assertTrue($this->platform->hasDoctrineTypeMappingFor('jsonb'));
+        self::assertEquals(Types::JSON, $this->platform->getDoctrineTypeMapping('jsonb'));
     }
 }
