+++ tests/Platforms/PostgreSQL100PlatformTest.php
@@ -7,7 +7,7 @@ namespace Doctrine\DBAL\Tests\Platforms;
 use Doctrine\DBAL\Platforms\AbstractPlatform;
 use Doctrine\DBAL\Platforms\PostgreSQL100Platform;
 
-class PostgreSQL100PlatformTest extends PostgreSQL94PlatformTest
+class PostgreSQL100PlatformTest extends PostgreSQLPlatformTest
 {
     public function createPlatform(): AbstractPlatform
     {
