+++ tests/Platforms/PostgreSQLPlatformTest.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Doctrine\DBAL\Tests\Platforms;
+
+use Doctrine\DBAL\Platforms\AbstractPlatform;
+use Doctrine\DBAL\Platforms\PostgreSQLPlatform;
+
+class PostgreSQLPlatformTest extends AbstractPostgreSQLPlatformTestCase
+{
+    public function createPlatform(): AbstractPlatform
+    {
+        return new PostgreSQLPlatform();
+    }
+}
